+++ src/components/memory/behavioral_l3_tile.cpp
@@ -33,7 +33,7 @@ void BehavioralL3Tile::reset() {
     }
 
     // Clear memory
-    std::fill(memory_.begin(), memory_.end(), 0);
+    std::fill(memory_.begin(), memory_.end(), static_cast<uint8_t>(0));
 
     // Reset statistics
     stats_.reset();
